day2/samples/my-app/backend/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
prometheus-client==0.17.1
gunicorn==21.2.0
pytest==7.4.2
pytest-cov==4.1.0
requests==2.31.0
